Refactor amp into three modules, separating clients/server better
